refactor(infra): (상품) 인프라스트럭처 계층 패키지 위치 변경 r
