--- src/main/java/kr/hhplus/be/server/product/domain/ProductRepository.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.product.domain;
-
-import java.util.Optional;
-
-public interface ProductRepository {
-    Optional<Product> findById(long productId);
-    Product insertOrUpdate(Product product);
-}
